backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
PyJWT>=2.8.0